Complete Chapter 15

+++ debug.h
@@ -4,6 +4,6 @@
 #include "chunk.h"
 
 void DisassembleChunk(Chunk *chunk, const char *name);
-int DisassembleInstruction(Chunk *chunk, int offset);
+int DisassembleInstruction(Chunk *chunk, size_t offset);
 
 #endif
